# v1/static/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Pi-Hole",
    "description": "The Pi-Hole® is a DNS sinkhole that protects your devices from unwanted content, without installing any client-side software.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/adlist": {
      "get": {
        "tags": [
          "AdLists"
        ],
        "summary": "Get all adlist",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/whitelist": {
      "get": {
        "tags": [
          "AdLists"
        ],
        "summary": "Get all whitelisted adlist (URL + Regex)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/blacklist": {
      "get": {
        "tags": [
          "AdLists"
        ],
        "summary": "Get all blacklisted adlist (URL + Regex)",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/adlist/group": {
      "get": {
        "tags": [
          "AdLists"
        ],
        "summary": "Get all adlist by group",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/clients": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Get all clients",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/clients/group": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Get all clients by group",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/domain/audit": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Get domain audit log",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/domains": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Get all domains",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/domains/group": {
      "get": {
        "tags": [
          "Domains"
        ],
        "summary": "Get all domains by group",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/gravity": {
      "get": {
        "tags": [
          "Gravity"
        ],
        "summary": "Get all gravity",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/groups": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "Get all groups",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/info": {
      "get": {
        "tags": [
          "Info"
        ],
        "summary": "Get all information",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/dnsinfo": {
      "get": {
        "tags": [
          "Info"
        ],
        "summary": "Get all dns information",
        "responses": {
          "200": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "405": {
            "description": "Not Allowed",
            "content": {}
          },
          "500": {
            "description": "Server Error",
            "content": {}
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "X-API-KEY",
        "in": "header"
      }
    }
  }
}